Check if .env exists and warn if HTTPS isn't being used
Chat API now checks if the `.env` exists. If not, it is created in either interactive or manual mode.

In Interactive mode, the user will be prompted for each field.

In manual mode, the .env is created with basic details and an explanation of what each field is for, plus default values.

"Connected to MongODB" message is also no longer repeated