Add CSS styling, modify server routes and refactor code
This large commit includes a lot of changes mainly focused on enhancing the CSS, modifying routing logic, and code refactoring for cleaner and more maintainable code. It introduces a new CSS style for documentations and modifies several server routes to improve server-client interactions.

It also includes several refactoring changes like enhancing the user-authentication middleware, improving response handling in the auth controller, and modifying the user data retrieval logic in several controllers for security purposes.

Moreover, unused code related to request statistics in app.js is removed and error handling is improved in this file as well as other controllers.